add client hook filter:videojs.options
closes #4086
