+++ client/src/root-helpers/index.ts
@@ -2,3 +2,4 @@ export * from './users'
 export * from './bytes'
 export * from './peertube-web-storage'
 export * from './utils'
+export * from './plugins-manager'
